refactor: centralise currency formatting using type level schema mappings

+++ src/main/java/io/github/shadrahman/fingraph/model/CategoryTotal.java
@@ -2,6 +2,6 @@ package io.github.shadrahman.fingraph.model;
 
 public record CategoryTotal(
         Category category,
-        Double amount
+        Long amount
 ) {
 }
